moved week8 assignments to week7
