+++ Week-7/detectLoopInLinkedList.js
